Changed insert to use SQL with scope identity.

+++ RandomPersonPicker.Data/Interfaces/Repositories/IPersonRepository.cs
@@ -8,7 +8,7 @@ namespace RandomPersonPicker.Data.Interfaces.Repositories
     {
         public Task<IEnumerable<Person>> Get();
         public Task<Person> Get(int personId);
-        public Task<long> Insert(Person person);
+        public Task<int> Insert(Person person);
         public Task<bool> Update(Person person);
     }
 }
